Update and refactor Settings component
