+++ src/sections/settings/SettingsPlanner.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const SettingsPlanner: React.FC = () => {
+  return (
+    <div className="px-0 md:px-6 mb-4 w-full md:max-w-1/3">
+      <h3 className="text-2xl font-light mb-6">Profile</h3>
+      <p className="text-sm text-gray-600">Planner settings will appear here.</p>
+    </div>
+  );
+};
+
+export default SettingsPlanner;
